tests/wpt/web-platform-tests/css/css-animations/animation-important-001.html

<!DOCTYPE html>
<title>Test !important declarations vs. animation effects</title>
<link rel="help" href="https://drafts.csswg.org/css-cascade/#cascade-origin">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  @keyframes bgcolor_animation {
    from { background-color: rgb(10, 10, 10); }
    to { background-color: rgb(20, 20, 20); }
  }
  @keyframes color_and_bg_animation {
    from { background-color: rgb(10, 10, 10); color: rgb(10, 10, 10); }
    to { background-color: rgb(20, 20, 20); color: rgb(20, 20, 20); }
  }
  a, div, ::before{
    animation-duration: 1000s;
    animation-delay: -500s;
    animation-timing-function: steps(2, end);
  }
  #target1 {
    animation-name: bgcolor_animation;
  }
  #target2 {
    background-color: rgb(0, 128, 0) !important;
    animation-name: bgcolor_animation;
  }
  #target3 {
    color: rgb(0, 128, 0) !important;
    animation-name: color_and_bg_animation;
  }
  #target4::before {
    color: rgb(0, 128, 0) !important;
    animation-name: color_and_bg_animation;
    content: " ";
  }
  #target5 {
    animation-name: color_and_bg_animation;
  }
  #target5:visited {
    color: rgb(0, 128, 0) !important;
  }
</style>
<div id="target1"></div>
<div id="target2"></div>
<div id="target3"></div>
<div id="target4"></div>
<a href="" id="target5"></a>
<script>
test(() => {
  assert_equals(getComputedStyle(target1).backgroundColor, 'rgb(15, 15, 15)');
}, 'Interpolated value is observable');

test(() => {
  assert_equals(getComputedStyle(target2).backgroundColor, 'rgb(0, 128, 0)');
}, 'Important rules override animations');

test(() => {
  assert_equals(getComputedStyle(target3).color, 'rgb(0, 128, 0)');
  assert_equals(getComputedStyle(target3).backgroundColor, 'rgb(15, 15, 15)');
}, 'Non-overriden interpolations are observable');

test(() => {
  assert_equals(getComputedStyle(target4, '::before').color, 'rgb(0, 128, 0)');
  assert_equals(getComputedStyle(target4, '::before').backgroundColor, 'rgb(15, 15, 15)');
}, 'Important rules override animations (::before)');

test(() => {
  assert_equals(getComputedStyle(target5).color, 'rgb(15, 15, 15)');
  assert_equals(getComputedStyle(target5).backgroundColor, 'rgb(15, 15, 15)');
}, 'Important rules do not override animations on :visited as seen from JS');

</script>
