third_party/blink/web_tests/virtual/do-not-hide-visited-color/external/wpt/css/css-animations/animation-important-001-expected.txt

This is a testharness.js-based test.
[FAIL] Important rules do not override animations on :visited as seen from JS
  assert_equals: expected "rgb(15, 15, 15)" but got "rgb(0, 128, 0)"
Harness: the test ran to completion.
